make center line processig default
